+++ src/renderer/language/en.js
@@ -29,6 +29,7 @@ module.exports = {
     menu_server_name: 'Server',
     session_clear_message: 'Clear Message',
     session_clear_message_confirm: 'Are you sure to clear the messages?',
-    session_no_more_message: 'There is no more messages.'
+    session_no_more_message: 'There is no more messages.',
+    session_without_client: 'No client connection.'
   }
 }