+++ src/renderer/views/Index.vue
@@ -1,18 +1,10 @@
 <template>
   <el-container>
     <Menu v-model="menu"></Menu>
-    <ClientSidebar
-      v-show="menu == CLIENT"
-      v-model="client"
-      @afterCreate="afterCreateClient"
-    ></ClientSidebar>
-    <ClientSession v-if="menu == CLIENT" ref="ClientSession" v-model="client"></ClientSession>
-    <ServerSidebar
-      v-show="menu == SERVER"
-      v-model="server"
-      @afterCreate="afterCreateServer"
-    ></ServerSidebar>
-    <ServerSession v-if="menu == SERVER" ref="ServerSession" v-model="server"></ServerSession>
+    <ClientSidebar v-show="menu == CLIENT" v-model="client" @afterCreate="afterCreateClient"></ClientSidebar>
+    <ClientSession v-show="menu == CLIENT" ref="ClientSession" v-model="client"></ClientSession>
+    <ServerSidebar v-show="menu == SERVER" v-model="server" @afterCreate="afterCreateServer"></ServerSidebar>
+    <ServerSession v-show="menu == SERVER" ref="ServerSession" v-model="server"></ServerSession>
   </el-container>
 </template>
 
@@ -35,7 +27,18 @@ export default {
       server: undefined
     }
   },
-  mounted() {},
+  watch: {
+    menu(newVal) {
+      this.$nextTick(() => {
+        if (newVal === SERVER) {
+          this.$refs.ServerSession.scrollBottom()
+        } else if (newVal === CLIENT) {
+          this.$refs.ClientSession.scrollBottom()
+        }
+      })
+    }
+  },
+  mounted() { },
   methods: {
     afterCreateClient(session, callback) {
       this.$nextTick(() => {